Feature/seab 1977 new logo (#121)
* Add Dockstore logo above ToC

Adding Dockstore logo to upper right corner (above Search & Table of Contents).

Two graphic files are checked in. The svg in case the file needs to be modified in the future.  The png for rendering.

* Add Dockstore logo above ToC

Adding Dockstore logo to upper right corner (above Search & Table of Contents).

Two graphic files are checked in. The svg in case the file needs to be modified in the future.  The png for rendering.

Also, updated conf.py to remove deprecated function call warning.  Changed app.add_stylesheet() to app.add_css_file().