Initial Dockstore Logo Usage Guidelines commit (#211)
* Initial Dockstore Logo Usage Guidelines commit

Dockstore Logo Usage Guidelines initial commit.

I was receiving an error message about "language = None" in the conf.py file.  Per the message, I changed it to "language = 'en'".

* Update no-discourse-topic-required.txt

List logo page in Discourse exclude list.

* Bring bullet list in compliance with style guide.

Updated the "Do Not" bullet list to comply with the style guide.

* Updated with feedback from PR

Misc. changes from the PR review.

* Further changes on the first paragraph.

Change the wording of the "About" paragraph.

* Misc PR requests

* Updated the zip file to include:
  * 1024 px width PNG variants of the logos & icons in addition to the existing sizes.
  * 32 px width variants of the icons
  * .eps variants of the files.

* Document changes
  * Move the link to the zip file above the logo table and adjusted the wording.
  * Deleted the "Do not" spacing requirement around the logo.

* Download link text change

A bit more text around the link to the zip file to help call attention to it.